requirements.txt
streamlit==1.28.0
fastapi>=0.100.0
uvicorn>=0.20.0
pandas==2.0.3
numpy==1.24.4
python-dotenv==1.0.0
boto3==1.28.62
great-expectations==0.17.15
pytest==7.4.2
pytest-cov==4.1.0
notebook==7.0.4
matplotlib==3.7.5
seaborn==0.13.0
scikit-learn==1.3.1
loguru==0.7.2
plotly==5.17.0
mlflow==2.8.1
kafka-python==2.0.2
pyjwt==2.8.0
psycopg2-binary==2.9.9
pyspark==3.5.3
delta-spark==3.0.0
